split dond out from do0d ..
