+++ qcodes/dataset/__init__.py
@@ -14,7 +14,10 @@ from .data_set_in_memory import load_from_netcdf
 from .data_set_protocol import DataSetProtocol, DataSetType
 from .database_extract_runs import extract_runs_into_db
 from .descriptions.param_spec import ParamSpec
-from .do_nd import AbstractSweep, ArraySweep, LinSweep, LogSweep, do0d, do1d, do2d, dond
+from .do_0d import do0d
+from .do_1d import do1d
+from .do_2d import do2d
+from .do_nd import AbstractSweep, ArraySweep, LinSweep, LogSweep, dond
 from .experiment_container import (
     experiments,
     load_experiment,
