+++ qcodes/dataset/do_nd.py
@@ -1,52 +1,41 @@
 from __future__ import annotations
 
 import logging
-import sys
 import time
 from abc import ABC, abstractmethod
-from contextlib import ExitStack, contextmanager
-from typing import Callable, Iterator, List, Mapping, Optional, Sequence, Tuple, Union
+from contextlib import ExitStack
+from typing import Mapping, Sequence
 
-import matplotlib.axes
-import matplotlib.colorbar
 import numpy as np
 from tqdm.auto import tqdm
 from typing_extensions import TypedDict
 
 from qcodes import config
-from qcodes.dataset.data_set_protocol import DataSetProtocol, res_type
+from qcodes.dataset.data_set_protocol import res_type
 from qcodes.dataset.descriptions.detect_shapes import detect_shape_of_measurement
 from qcodes.dataset.descriptions.versioning.rundescribertypes import Shapes
 from qcodes.dataset.experiment_container import Experiment
 from qcodes.dataset.measurements import Measurement
-from qcodes.dataset.plotting import plot_and_save_image
 from qcodes.parameters import ParameterBase
 
+from .do_nd_utils import (
+    ActionsT,
+    AxesTupleListWithDataSet,
+    BreakConditionInterrupt,
+    BreakConditionT,
+    MultiAxesTupleListWithDataSet,
+    ParamMeasT,
+    _catch_interrupts,
+    _handle_plotting,
+    _register_actions,
+    _register_parameters,
+    _set_write_period,
+)
 from .threading import (
     SequentialParamsCaller,
     ThreadPoolParamsCaller,
     process_params_meas,
 )
-
-ActionsT = Sequence[Callable[[], None]]
-BreakConditionT = Callable[[], bool]
-
-ParamMeasT = Union[ParameterBase, Callable[[], None]]
-
-AxesTuple = Tuple[matplotlib.axes.Axes, matplotlib.colorbar.Colorbar]
-AxesTupleList = Tuple[
-    List[matplotlib.axes.Axes], List[Optional[matplotlib.colorbar.Colorbar]]
-]
-AxesTupleListWithDataSet = Tuple[
-    DataSetProtocol,
-    List[matplotlib.axes.Axes],
-    List[Optional[matplotlib.colorbar.Colorbar]],
-]
-MultiAxesTupleListWithDataSet = Tuple[
-    Tuple[DataSetProtocol, ...],
-    Tuple[List[matplotlib.axes.Axes], ...],
-    Tuple[List[Optional[matplotlib.colorbar.Colorbar]], ...],
-]
 
 LOG = logging.getLogger(__name__)
 
@@ -59,425 +48,6 @@ class ParameterGroup(TypedDict):
 
 class UnsafeThreadingException(Exception):
     pass
-
-
-class BreakConditionInterrupt(Exception):
-    pass
-
-
-MeasInterruptT = Union[KeyboardInterrupt, BreakConditionInterrupt, None]
-
-
-def _register_parameters(
-    meas: Measurement,
-    param_meas: Sequence[ParamMeasT],
-    setpoints: Sequence[ParameterBase] | None = None,
-    shapes: Shapes = None,
-) -> None:
-    for parameter in param_meas:
-        if isinstance(parameter, ParameterBase):
-            meas.register_parameter(parameter, setpoints=setpoints)
-    meas.set_shapes(shapes=shapes)
-
-
-def _register_actions(
-    meas: Measurement, enter_actions: ActionsT, exit_actions: ActionsT
-) -> None:
-    for action in enter_actions:
-        # this omits the possibility of passing
-        # argument to enter and exit actions.
-        # Do we want that?
-        meas.add_before_run(action, ())
-    for action in exit_actions:
-        meas.add_after_run(action, ())
-
-
-def _set_write_period(meas: Measurement, write_period: float | None = None) -> None:
-    if write_period is not None:
-        meas.write_period = write_period
-
-
-@contextmanager
-def _catch_interrupts() -> Iterator[Callable[[], MeasInterruptT]]:
-    interrupt_exception = None
-
-    def get_interrupt_exception() -> MeasInterruptT:
-        nonlocal interrupt_exception
-        return interrupt_exception
-
-    try:
-        yield get_interrupt_exception
-    except (KeyboardInterrupt, BreakConditionInterrupt) as e:
-        interrupt_exception = e
-
-
-def do0d(
-    *param_meas: ParamMeasT,
-    write_period: float | None = None,
-    measurement_name: str = "",
-    exp: Experiment | None = None,
-    do_plot: bool | None = None,
-    use_threads: bool | None = None,
-    log_info: str | None = None,
-) -> AxesTupleListWithDataSet:
-    """
-    Perform a measurement of a single parameter. This is probably most
-    useful for an ArrayParameter that already returns an array of data points
-
-    Args:
-        *param_meas: Parameter(s) to measure at each step or functions that
-          will be called at each step. The function should take no arguments.
-          The parameters and functions are called in the order they are
-          supplied.
-        write_period: The time after which the data is actually written to the
-            database.
-        measurement_name: Name of the measurement. This will be passed down to
-            the dataset produced by the measurement. If not given, a default
-            value of 'results' is used for the dataset.
-        exp: The experiment to use for this measurement.
-        do_plot: should png and pdf versions of the images be saved after the
-            run. If None the setting will be read from ``qcodesrc.json``
-        use_threads: If True measurements from each instrument will be done on
-            separate threads. If you are measuring from several instruments
-            this may give a significant speedup.
-        log_info: Message that is logged during the measurement. If None a default
-            message is used.
-
-    Returns:
-        The QCoDeS dataset.
-    """
-    if do_plot is None:
-        do_plot = config.dataset.dond_plot
-    meas = Measurement(name=measurement_name, exp=exp)
-    if log_info is not None:
-        meas._extra_log_info = log_info
-    else:
-        meas._extra_log_info = "Using 'qcodes.dataset.do0d'"
-
-    measured_parameters = tuple(
-        param for param in param_meas if isinstance(param, ParameterBase)
-    )
-
-    try:
-        shapes: Shapes = detect_shape_of_measurement(
-            measured_parameters,
-        )
-    except TypeError:
-        LOG.exception(
-            f"Could not detect shape of {measured_parameters} "
-            f"falling back to unknown shape."
-        )
-        shapes = None
-
-    _register_parameters(meas, param_meas, shapes=shapes)
-    _set_write_period(meas, write_period)
-
-    with meas.run() as datasaver:
-        datasaver.add_result(*process_params_meas(param_meas, use_threads=use_threads))
-        dataset = datasaver.dataset
-
-    return _handle_plotting(dataset, do_plot)
-
-
-def do1d(
-    param_set: ParameterBase,
-    start: float,
-    stop: float,
-    num_points: int,
-    delay: float,
-    *param_meas: ParamMeasT,
-    enter_actions: ActionsT = (),
-    exit_actions: ActionsT = (),
-    write_period: float | None = None,
-    measurement_name: str = "",
-    exp: Experiment | None = None,
-    do_plot: bool | None = None,
-    use_threads: bool | None = None,
-    additional_setpoints: Sequence[ParameterBase] = tuple(),
-    show_progress: bool | None = None,
-    log_info: str | None = None,
-    break_condition: BreakConditionT | None = None,
-) -> AxesTupleListWithDataSet:
-    """
-    Perform a 1D scan of ``param_set`` from ``start`` to ``stop`` in
-    ``num_points`` measuring param_meas at each step. In case param_meas is
-    an ArrayParameter this is effectively a 2d scan.
-
-    Args:
-        param_set: The QCoDeS parameter to sweep over
-        start: Starting point of sweep
-        stop: End point of sweep
-        num_points: Number of points in sweep
-        delay: Delay after setting parameter before measurement is performed
-        param_meas: Parameter(s) to measure at each step or functions that
-          will be called at each step. The function should take no arguments.
-          The parameters and functions are called in the order they are
-          supplied.
-        enter_actions: A list of functions taking no arguments that will be
-            called before the measurements start
-        exit_actions: A list of functions taking no arguments that will be
-            called after the measurements ends
-        write_period: The time after which the data is actually written to the
-            database.
-        additional_setpoints: A list of setpoint parameters to be registered in
-            the measurement but not scanned.
-        measurement_name: Name of the measurement. This will be passed down to
-            the dataset produced by the measurement. If not given, a default
-            value of 'results' is used for the dataset.
-        exp: The experiment to use for this measurement.
-        do_plot: should png and pdf versions of the images be saved after the
-            run. If None the setting will be read from ``qcodesrc.json``
-        use_threads: If True measurements from each instrument will be done on
-            separate threads. If you are measuring from several instruments
-            this may give a significant speedup.
-        show_progress: should a progress bar be displayed during the
-            measurement. If None the setting will be read from ``qcodesrc.json``
-        log_info: Message that is logged during the measurement. If None a default
-            message is used.
-        break_condition: Callable that takes no arguments. If returned True,
-            measurement is interrupted.
-
-    Returns:
-        The QCoDeS dataset.
-    """
-    if do_plot is None:
-        do_plot = config.dataset.dond_plot
-    if show_progress is None:
-        show_progress = config.dataset.dond_show_progress
-
-    meas = Measurement(name=measurement_name, exp=exp)
-    if log_info is not None:
-        meas._extra_log_info = log_info
-    else:
-        meas._extra_log_info = "Using 'qcodes.dataset.do1d'"
-
-    all_setpoint_params = (param_set,) + tuple(s for s in additional_setpoints)
-
-    measured_parameters = tuple(
-        param for param in param_meas if isinstance(param, ParameterBase)
-    )
-    try:
-        loop_shape = (num_points,) + tuple(1 for _ in additional_setpoints)
-        shapes: Shapes = detect_shape_of_measurement(measured_parameters, loop_shape)
-    except TypeError:
-        LOG.exception(
-            f"Could not detect shape of {measured_parameters} "
-            f"falling back to unknown shape."
-        )
-        shapes = None
-
-    _register_parameters(meas, all_setpoint_params)
-    _register_parameters(meas, param_meas, setpoints=all_setpoint_params, shapes=shapes)
-    _set_write_period(meas, write_period)
-    _register_actions(meas, enter_actions, exit_actions)
-
-    if use_threads is None:
-        use_threads = config.dataset.use_threads
-
-    param_meas_caller = (
-        ThreadPoolParamsCaller(*param_meas)
-        if use_threads
-        else SequentialParamsCaller(*param_meas)
-    )
-
-    # do1D enforces a simple relationship between measured parameters
-    # and set parameters. For anything more complicated this should be
-    # reimplemented from scratch
-    with _catch_interrupts() as interrupted, meas.run() as datasaver, param_meas_caller as call_param_meas:
-        dataset = datasaver.dataset
-        additional_setpoints_data = process_params_meas(additional_setpoints)
-        setpoints = np.linspace(start, stop, num_points)
-
-        # flush to prevent unflushed print's to visually interrupt tqdm bar
-        # updates
-        sys.stdout.flush()
-        sys.stderr.flush()
-
-        for set_point in tqdm(setpoints, disable=not show_progress):
-            param_set.set(set_point)
-            time.sleep(delay)
-            datasaver.add_result(
-                (param_set, set_point), *call_param_meas(), *additional_setpoints_data
-            )
-
-            if callable(break_condition):
-                if break_condition():
-                    raise BreakConditionInterrupt("Break condition was met.")
-
-    return _handle_plotting(dataset, do_plot, interrupted())
-
-
-def do2d(
-    param_set1: ParameterBase,
-    start1: float,
-    stop1: float,
-    num_points1: int,
-    delay1: float,
-    param_set2: ParameterBase,
-    start2: float,
-    stop2: float,
-    num_points2: int,
-    delay2: float,
-    *param_meas: ParamMeasT,
-    set_before_sweep: bool | None = True,
-    enter_actions: ActionsT = (),
-    exit_actions: ActionsT = (),
-    before_inner_actions: ActionsT = (),
-    after_inner_actions: ActionsT = (),
-    write_period: float | None = None,
-    measurement_name: str = "",
-    exp: Experiment | None = None,
-    flush_columns: bool = False,
-    do_plot: bool | None = None,
-    use_threads: bool | None = None,
-    additional_setpoints: Sequence[ParameterBase] = tuple(),
-    show_progress: bool | None = None,
-    log_info: str | None = None,
-    break_condition: BreakConditionT | None = None,
-) -> AxesTupleListWithDataSet:
-    """
-    Perform a 1D scan of ``param_set1`` from ``start1`` to ``stop1`` in
-    ``num_points1`` and ``param_set2`` from ``start2`` to ``stop2`` in
-    ``num_points2`` measuring param_meas at each step.
-
-    Args:
-        param_set1: The QCoDeS parameter to sweep over in the outer loop
-        start1: Starting point of sweep in outer loop
-        stop1: End point of sweep in the outer loop
-        num_points1: Number of points to measure in the outer loop
-        delay1: Delay after setting parameter in the outer loop
-        param_set2: The QCoDeS parameter to sweep over in the inner loop
-        start2: Starting point of sweep in inner loop
-        stop2: End point of sweep in the inner loop
-        num_points2: Number of points to measure in the inner loop
-        delay2: Delay after setting parameter before measurement is performed
-        param_meas: Parameter(s) to measure at each step or functions that
-          will be called at each step. The function should take no arguments.
-          The parameters and functions are called in the order they are
-          supplied.
-        set_before_sweep: if True the outer parameter is set to its first value
-            before the inner parameter is swept to its next value.
-        enter_actions: A list of functions taking no arguments that will be
-            called before the measurements start
-        exit_actions: A list of functions taking no arguments that will be
-            called after the measurements ends
-        before_inner_actions: Actions executed before each run of the inner loop
-        after_inner_actions: Actions executed after each run of the inner loop
-        write_period: The time after which the data is actually written to the
-            database.
-        measurement_name: Name of the measurement. This will be passed down to
-            the dataset produced by the measurement. If not given, a default
-            value of 'results' is used for the dataset.
-        exp: The experiment to use for this measurement.
-        flush_columns: The data is written after a column is finished
-            independent of the passed time and write period.
-        additional_setpoints: A list of setpoint parameters to be registered in
-            the measurement but not scanned.
-        do_plot: should png and pdf versions of the images be saved after the
-            run. If None the setting will be read from ``qcodesrc.json``
-        use_threads: If True measurements from each instrument will be done on
-            separate threads. If you are measuring from several instruments
-            this may give a significant speedup.
-        show_progress: should a progress bar be displayed during the
-            measurement. If None the setting will be read from ``qcodesrc.json``
-        log_info: Message that is logged during the measurement. If None a default
-            message is used.
-        break_condition: Callable that takes no arguments. If returned True,
-            measurement is interrupted.
-
-    Returns:
-        The QCoDeS dataset.
-    """
-
-    if do_plot is None:
-        do_plot = config.dataset.dond_plot
-    if show_progress is None:
-        show_progress = config.dataset.dond_show_progress
-
-    meas = Measurement(name=measurement_name, exp=exp)
-    if log_info is not None:
-        meas._extra_log_info = log_info
-    else:
-        meas._extra_log_info = "Using 'qcodes.dataset.do2d'"
-    all_setpoint_params = (
-        param_set1,
-        param_set2,
-    ) + tuple(s for s in additional_setpoints)
-
-    measured_parameters = tuple(
-        param for param in param_meas if isinstance(param, ParameterBase)
-    )
-
-    try:
-        loop_shape = (num_points1, num_points2) + tuple(1 for _ in additional_setpoints)
-        shapes: Shapes = detect_shape_of_measurement(measured_parameters, loop_shape)
-    except TypeError:
-        LOG.exception(
-            f"Could not detect shape of {measured_parameters} "
-            f"falling back to unknown shape."
-        )
-        shapes = None
-
-    _register_parameters(meas, all_setpoint_params)
-    _register_parameters(meas, param_meas, setpoints=all_setpoint_params, shapes=shapes)
-    _set_write_period(meas, write_period)
-    _register_actions(meas, enter_actions, exit_actions)
-
-    if use_threads is None:
-        use_threads = config.dataset.use_threads
-
-    param_meas_caller = (
-        ThreadPoolParamsCaller(*param_meas)
-        if use_threads
-        else SequentialParamsCaller(*param_meas)
-    )
-
-    with _catch_interrupts() as interrupted, meas.run() as datasaver, param_meas_caller as call_param_meas:
-        dataset = datasaver.dataset
-        additional_setpoints_data = process_params_meas(additional_setpoints)
-        setpoints1 = np.linspace(start1, stop1, num_points1)
-        for set_point1 in tqdm(setpoints1, disable=not show_progress):
-            if set_before_sweep:
-                param_set2.set(start2)
-
-            param_set1.set(set_point1)
-
-            for action in before_inner_actions:
-                action()
-
-            time.sleep(delay1)
-
-            setpoints2 = np.linspace(start2, stop2, num_points2)
-
-            # flush to prevent unflushed print's to visually interrupt tqdm bar
-            # updates
-            sys.stdout.flush()
-            sys.stderr.flush()
-            for set_point2 in tqdm(setpoints2, disable=not show_progress, leave=False):
-                # skip first inner set point if `set_before_sweep`
-                if set_point2 == start2 and set_before_sweep:
-                    pass
-                else:
-                    param_set2.set(set_point2)
-                    time.sleep(delay2)
-
-                datasaver.add_result(
-                    (param_set1, set_point1),
-                    (param_set2, set_point2),
-                    *call_param_meas(),
-                    *additional_setpoints_data,
-                )
-
-                if callable(break_condition):
-                    if break_condition():
-                        raise BreakConditionInterrupt("Break condition was met.")
-
-            for action in after_inner_actions:
-                action()
-            if flush_columns:
-                datasaver.flush_data_to_database()
-
-    return _handle_plotting(dataset, do_plot, interrupted())
 
 
 class AbstractSweep(ABC):
@@ -1092,28 +662,3 @@ def _extract_paramters_by_type_and_group(
             }
             grouped_parameters[f"group_{index}"] = pg
     return tuple(measured_all), grouped_parameters, tuple(measured_parameters)
-
-
-def _handle_plotting(
-    data: DataSetProtocol,
-    do_plot: bool = True,
-    interrupted: MeasInterruptT = None,
-) -> AxesTupleListWithDataSet:
-    """
-    Save the plots created by datasaver as pdf and png
-
-    Args:
-        datasaver: a measurement datasaver that contains a dataset to be saved
-            as plot.
-            :param do_plot:
-
-    """
-    if do_plot:
-        res = plot_and_save_image(data)
-    else:
-        res = data, [None], [None]
-
-    if interrupted:
-        raise interrupted
-
-    return res
