+++ qcodes/utils/dataset/doNd.py
@@ -1,36 +1,38 @@
 import warnings
 
+from qcodes.dataset.do_0d import do0d
+from qcodes.dataset.do_1d import do1d
+from qcodes.dataset.do_2d import do2d
 from qcodes.dataset.do_nd import (
     AbstractSweep,
-    ActionsT,
     ArraySweep,
+    LinSweep,
+    LogSweep,
+    ParameterGroup,
+    UnsafeThreadingException,
+    _conditional_parameter_set,
+    _create_measurements,
+    _extract_paramters_by_type_and_group,
+    _make_nested_setpoints,
+    _parse_dond_arguments,
+    _select_active_actions_delays,
+    dond,
+)
+from qcodes.dataset.do_nd_utils import (
+    ActionsT,
     AxesTuple,
     AxesTupleList,
     AxesTupleListWithDataSet,
     BreakConditionInterrupt,
     BreakConditionT,
-    LinSweep,
-    LogSweep,
     MeasInterruptT,
     MultiAxesTupleListWithDataSet,
-    ParameterGroup,
     ParamMeasT,
-    UnsafeThreadingException,
     _catch_interrupts,
-    _conditional_parameter_set,
-    _create_measurements,
-    _extract_paramters_by_type_and_group,
     _handle_plotting,
-    _make_nested_setpoints,
-    _parse_dond_arguments,
     _register_actions,
     _register_parameters,
-    _select_active_actions_delays,
     _set_write_period,
-    do0d,
-    do1d,
-    do2d,
-    dond,
 )
 from qcodes.dataset.plotting import plot_and_save_image as plot
